requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
django-filter>=23.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
dj-database-url>=2.1.0